filesystem/ConfigHelper: relocate to youtils
Part of https://github.com/openvstorage/volumedriver/issues/53
