+++ src/filesystem/LocalPythonClient.cpp
@@ -13,7 +13,6 @@
 // Open vStorage is distributed in the hope that it will be useful,
 // but WITHOUT ANY WARRANTY of any kind.
 
-#include "ConfigFetcher.h"
 #include "ConfigHelper.h"
 #include "FileSystem.h"
 #include "LocalPythonClient.h"
