+++ src/filesystem/Makefile.am
@@ -64,7 +64,6 @@ libvolumedriverfs_la_SOURCES = \
 		ClusterNodeConfig.cpp \
 		ClusterNodeStatus.cpp \
 		ClusterRegistry.cpp \
-		ConfigFetcher.cpp \
 		ConfigHelper.cpp \
 		DataPoints.cpp \
 		DirectoryEntry.cpp \
